pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-test-automation-mcp"
version = "0.1.0"
description = "Autonomous web test agent using MCP and Chrome DevTools"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "mcp>=1.10.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "tenacity>=8.5.0",
  "rich>=13.7.1",
  "easyocr>=1.7.2",
  "opencv-python>=4.10.0.84",
  "Pillow>=10.4.0",
  "numpy>=1.26.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
